Make EntityType return integers for dimensions
The constructor takes in integers, so there is no point in returning
anything of higher precision.

+++ core/src/dev/theavid/periculum/entities/EntityType.java
@@ -64,11 +64,11 @@ public enum EntityType {
 		return collidable;
 	}
 
-	public double getWidth() {
+	public int getWidth() {
 		return width;
 	}
 
-	public double getHeight() {
+	public int getHeight() {
 		return height;
 	}
 }
